Move label classes to seperate package
